requirements.txt
blinker==1.9.0
click==8.3.3
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.3
fonttools==4.62.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.9
numpy==2.4.4
packaging==26.2
pillow==12.2.0
pycryptodome==3.23.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
secretsharing==0.2.6
shamir-mnemonic==0.3.0
six==1.17.0
Werkzeug==3.1.8
